+++ Extension/src/background/api/filtering-log.ts
@@ -30,7 +30,7 @@ import {
 } from '@adguard/tswebextension';
 
 import { AntiBannerFiltersId } from '../../common/constants';
-import { Log } from '../../common/log';
+import { logger } from '../../common/logger';
 import { translator } from '../../common/translators/translator';
 import { listeners } from '../notifier';
 import { Engine } from '../engine';
@@ -143,13 +143,13 @@ export class FilteringLogApi {
         try {
             Engine.api.setDebugScriptlets(true);
         } catch (e) {
-            Log.error('Failed to enable `verbose scriptlets logging` option', e);
+            logger.error('Failed to enable `verbose scriptlets logging` option', e);
         }
 
         try {
             Engine.api.setCollectHitStats(true);
         } catch (e) {
-            Log.error('Failed to enable `collect hit stats` option', e);
+            logger.error('Failed to enable `collect hit stats` option', e);
         }
     }
 
@@ -167,14 +167,14 @@ export class FilteringLogApi {
             try {
                 Engine.api.setDebugScriptlets(false);
             } catch (e) {
-                Log.error('Failed to disable `verbose scriptlets logging` option', e);
+                logger.error('Failed to disable `verbose scriptlets logging` option', e);
             }
 
             if (settingsStorage.get(SettingOption.DisableCollectHits)) {
                 try {
                     Engine.api.setCollectHitStats(false);
                 } catch (e) {
-                    Log.error('Failed to disable `collect hit stats` option', e);
+                    logger.error('Failed to disable `collect hit stats` option', e);
                 }
             }
         }