+++ Extension/src/background/api/filters/common.ts
@@ -18,7 +18,7 @@
 import browser from 'webextension-polyfill';
 
 import { BrowserUtils } from '../../utils/browser-utils';
-import { Log } from '../../../common/log';
+import { logger } from '../../../common/logger';
 import { UserAgent } from '../../../common/user-agent';
 import { SettingOption, RegularFilterMetadata } from '../../schema';
 import { AntiBannerFiltersId } from '../../../common/constants';
@@ -91,7 +91,7 @@ export class CommonFilterApi {
     public static async updateFilter(
         filterUpdateOptions: FilterUpdateOptions,
     ): Promise<RegularFilterMetadata | null> {
-        Log.info(`Update filter ${filterUpdateOptions.filterId}`);
+        logger.info(`Update filter ${filterUpdateOptions.filterId}`);
 
         // We do not have to check metadata for the filters which do not update with force, because
         // they even do not trigger metadata update.
@@ -99,24 +99,24 @@ export class CommonFilterApi {
             const filterMetadata = CommonFilterApi.getFilterMetadata(filterUpdateOptions.filterId);
 
             if (!filterMetadata) {
-                Log.error(`Cannot find filter ${filterUpdateOptions.filterId} metadata`);
+                logger.error(`Cannot find filter ${filterUpdateOptions.filterId} metadata`);
                 return null;
             }
 
             if (!CommonFilterApi.isFilterNeedUpdate(filterMetadata)) {
-                Log.info(`Filter ${filterUpdateOptions.filterId} is already updated`);
+                logger.info(`Filter ${filterUpdateOptions.filterId} is already updated`);
                 return null;
             }
         }
 
-        Log.info(`Filter ${filterUpdateOptions.filterId} is need to updated`);
+        logger.info(`Filter ${filterUpdateOptions.filterId} is need to updated`);
 
         try {
             const filterMetadata = await CommonFilterApi.loadFilterRulesFromBackend(filterUpdateOptions, true);
-            Log.info(`Filter ${filterUpdateOptions.filterId} updated successfully`);
+            logger.info(`Filter ${filterUpdateOptions.filterId} updated successfully`);
             return filterMetadata;
         } catch (e) {
-            Log.error(e);
+            logger.error(e);
             return null;
         }
     }
@@ -268,7 +268,7 @@ export class CommonFilterApi {
      * @returns True, if filter update is required, else returns false.
      */
     private static isFilterNeedUpdate(filterMetadata: RegularFilterMetadata): boolean {
-        Log.info(`Check if filter ${filterMetadata.filterId} need to update`);
+        logger.info(`Check if filter ${filterMetadata.filterId} need to update`);
 
         const filterVersion = filterVersionStorage.get(filterMetadata.filterId);