+++ Extension/src/background/services/ui/main.ts
@@ -24,7 +24,7 @@ import {
     tabsApi as tsWebExtTabApi,
 } from '@adguard/tswebextension';
 
-import { Log } from '../../../common/log';
+import { logger } from '../../../common/logger';
 import { messageHandler } from '../../message-handler';
 import {
     MessageType,
@@ -153,7 +153,7 @@ export class UiService {
         if (activeTab?.url) {
             await PagesApi.openAbusePage(activeTab.url, ForwardFrom.ContextMenu);
         } else {
-            Log.warn('Cannot open abuse page for active tab');
+            logger.warn('Cannot open abuse page for active tab');
         }
     }
 
@@ -178,7 +178,7 @@ export class UiService {
         if (activeTab?.url) {
             await PagesApi.openSiteReportPage(activeTab.url, ForwardFrom.ContextMenu);
         } else {
-            Log.warn('Cannot open site report page for active tab');
+            logger.warn('Cannot open site report page for active tab');
         }
     }